requirements.txt
Flask
Flask-Cors
requests
firebase-admin
python-dotenv
sentence-transformers
chromadb
google-cloud-aiplatform